Enum matrix_qrcode::QrVerificationData

pub enum QrVerificationData {
    Verification(VerificationData),
    SelfVerification(SelfVerificationData),
    SelfVerificationNoMasterKey(SelfVerificationNoMasterKey),
}

An enum representing the different modes a QR verification can be in.

Variants

Verification(VerificationData)

The QR verification is verifying another user

SelfVerification(SelfVerificationData)

The QR verification is self-verifying and the current device trusts or owns the master key

SelfVerificationNoMasterKey(SelfVerificationNoMasterKey)

The QR verification is self-verifying in which the current device does not yet trust the master key

Implementations

impl QrVerificationData

pub fn from_image(image: DynamicImage) -> Result<Self, DecodingError>

This is supported on decode_image only.

Decode and parse an image of a QR code into a QrVerificationData

The image will be converted into a grey scale image before decoding is attempted

Arguments

image - The image containing the QR code.

Example

use image;

let image = image::open("/path/to/my/image.png").unwrap();
let result = QrVerificationData::from_image(image)?;

pub fn from_luma(
image: ImageBuffer<Luma<u8>, Vec<u8>>
) -> Result<Self, DecodingError>

This is supported on decode_image only.

Decode and parse an grey scale image of a QR code into a QrVerificationData

Arguments

image - The grey scale image containing the QR code.

Example

use image;

let image = image::open("/path/to/my/image.png").unwrap();
let image = image.to_luma8();
let result = QrVerificationData::from_luma(image)?;

pub fn from_bytes(bytes: impl AsRef<[u8]>) -> Result<Self, DecodingError>

Parse the decoded payload of a QR code in byte slice form as a QrVerificationData

This method is useful if you would like to do your own custom QR code decoding.

Arguments

bytes - The raw bytes of a decoded QR code.

Example

let data = b"MATRIX\
             \x02\x02\x00\x07\
             FLOW_ID\
             AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA\
             BBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBB\
             SHARED_SECRET";

let result = QrVerificationData::from_bytes(data)?;

pub fn to_qr_code(&self) -> Result<QrCode, EncodingError>

Encode the QrVerificationData into a QrCode.

This method turns the QrVerificationData into a QR code that can be rendered and presented to be scanned.

The encoding can fail if the data doesn’t fit into a QR code or if the identity keys that should be encoded into the QR code are not valid base64.

Example

let data = b"MATRIX\
             \x02\x02\x00\x07\
             FLOW_ID\
             AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA\
             BBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBB\
             SHARED_SECRET";

let result = QrVerificationData::from_bytes(data)?;
let encoded = result.to_qr_code().unwrap();

pub fn to_bytes(&self) -> Result<Vec<u8>, EncodingError>

Encode the QrVerificationData into a vector of bytes that can be encoded as a QR code.

The encoding can fail if the identity keys that should be encoded are not valid base64.

Example

let data = b"MATRIX\
             \x02\x02\x00\x07\
             FLOW_ID\
             AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA\
             BBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBB\
             SHARED_SECRET";

let result = QrVerificationData::from_bytes(data)?;
let encoded = result.to_bytes().unwrap();

assert_eq!(data.as_ref(), encoded.as_slice());

pub fn flow_id(&self) -> &str

Get the flow id for this QrVerificationData.

This represents the ID as a string even if it is a EventId.

pub fn first_key(&self) -> &str

Get the first key of this QrVerificationData.

pub fn second_key(&self) -> &str

Get the second key of this QrVerificationData.

pub fn secret(&self) -> &str

Get the secret of this QrVerificationData.