Crate matrix_qrcode

This crate implements methods to parse and generate QR codes that are used for interactive verification in Matrix.

It implements the QR format defined in the Matrix spec.

use image;

let image = image::open("/path/to/my/image.png").unwrap();
let result = QrVerificationData::from_image(image)?;

Re-exports

pub use image;

pub use qrcode;

pub use rqrr;

Structs

SelfVerificationData

The non-encoded data for the second mode of QR code verification.

SelfVerificationNoMasterKey

The non-encoded data for the third mode of QR code verification.

VerificationData

The non-encoded data for the first mode of QR code verification.

Enums

DecodingError

Error type describing errors that happen while QR data is being decoded.

EncodingError

Error type describing errors that happen while QR data is being encoded.

QrVerificationData

An enum representing the different modes a QR verification can be in.